bug fix: change type of courseNum in database, isregisterd,unregister
change kdamCourses to sorted set and acording to thah change inialize in database

+++ spl-net/src/main/java/bgu/spl/net/api/Messages/IsRegisteredMessage.java
@@ -4,8 +4,8 @@ import bgu.spl.net.api.Message;
 import bgu.spl.net.impl.RegistrationSystem.*;
 
 public class IsRegisteredMessage implements Message {
-    int courseNum;
-    public IsRegisteredMessage(int courseNum){
+    Short courseNum;
+    public IsRegisteredMessage(Short courseNum){
         super();
         this.courseNum=courseNum;
     }
